refactor: 동적 리소스 컨트롤러 분리
