+++ tomcat/src/main/java/com/techcourse/controller/IndexController.java
@@ -1,12 +1,12 @@
 package com.techcourse.controller;
 
-import org.apache.catalina.controller.AbstractController;
+import org.apache.catalina.controller.DynamicSourceController;
 import org.apache.coyote.http.ContentType;
 import org.apache.coyote.http.Header;
 import org.apache.coyote.http.request.HttpRequest;
 import org.apache.coyote.http.response.HttpResponse;
 
-public class IndexController extends AbstractController {
+public class IndexController extends DynamicSourceController {
 
     private static final String HELLO_PAGE = "/hello.html";
