+++ tomcat/src/main/java/com/techcourse/controller/RegisterController.java
@@ -2,11 +2,11 @@ package com.techcourse.controller;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
-import org.apache.catalina.controller.AbstractController;
+import org.apache.catalina.controller.DynamicSourceController;
 import org.apache.coyote.http.request.HttpRequest;
 import org.apache.coyote.http.response.HttpResponse;
 
-public class RegisterController extends AbstractController {
+public class RegisterController extends DynamicSourceController {
 
     private static final String LOGIN_PAGE = "/login.html";
     private static final String REGISTER_PAGE = "/register.html";
